cleanup & generalize contract msg
