+++ local-interchaintest/examples/icq_test.rs
@@ -1,5 +1,8 @@
 use local_interchaintest::utils::{
-    icq::{generate_icq_relayer_config, start_icq_relayer},
+    icq::{
+        generate_icq_relayer_config, query_catchall_logs, query_raw_result,
+        register_kvq_balances_query, start_icq_relayer, try_parse_storage_value,
+    },
     GAS_FLAGS, LOGS_FILE_PATH, VALENCE_ARTIFACTS_PATH,
 };
 use localic_std::{
@@ -8,7 +11,13 @@ use localic_std::{
     types::TransactionResponse,
 };
 use log::info;
-use neutron_sdk::bindings::types::{InterchainQueryResult, StorageValue};
+use neutron_sdk::{
+    bindings::types::{InterchainQueryResult, StorageValue},
+    interchain_queries::{
+        helpers::decode_and_convert,
+        v047::{helpers::create_account_denom_balance_key, types::BANK_STORE_KEY},
+    },
+};
 use serde_json::Value;
 use std::{env, error::Error, time::Duration};
 use valence_test_icq_lib::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
@@ -18,6 +27,13 @@ use localic_utils::{
     LOCAL_IC_API_URL, NEUTRON_CHAIN_NAME, OSMOSIS_CHAIN_ADMIN_ADDR, OSMOSIS_CHAIN_DENOM,
     OSMOSIS_CHAIN_NAME,
 };
+
+// KeyNextGlobalPoolId defines key to store the next Pool ID to be used.
+pub const NEXT_GLOBAL_POOL_ID_KEY: u8 = 0x01;
+pub const PREFIX_POOLS_KEY: u8 = 0x02;
+pub const TOTAL_LIQUIDITY_KEY: u8 = 0x03;
+pub const PREFIX_MIGRATION_INFO_BALANCER_POOL_KEY: u8 = 0x04;
+pub const GAMM_STORE_KEY: &str = "gamm";
 
 fn main() -> Result<(), Box<dyn Error>> {
     env_logger::init();
@@ -83,35 +99,42 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     info!("icq test lib: {}", icq_test_lib.address);
 
-    let icq_registration_response = register_icq_balances_query(
-        &test_ctx,
-        icq_test_lib.address.to_string(),
-        OSMOSIS_CHAIN_NAME.to_string(),
-        OSMOSIS_CHAIN_ADMIN_ADDR.to_string(),
-        vec![OSMOSIS_CHAIN_DENOM.to_string()],
-    )?;
+    // let icq_registration_response = register_icq_balances_query(
+    //     &test_ctx,
+    //     icq_test_lib.address.to_string(),
+    //     OSMOSIS_CHAIN_NAME.to_string(),
+    //     OSMOSIS_CHAIN_ADMIN_ADDR.to_string(),
+    //     vec![OSMOSIS_CHAIN_DENOM.to_string()],
+    // )?;
 
-    info!("icq registration response: {:?}", icq_registration_response);
+    // info!("icq registration response: {:?}", icq_registration_response);
 
-    for _ in 0..10 {
-        let response_value =
-            query_balance_query_id(&test_ctx, icq_test_lib.address.to_string(), 1)?;
+    // for _ in 0..10 {
+    //     let response_value =
+    //         query_balance_query_id(&test_ctx, icq_test_lib.address.to_string(), 1)?;
 
-        info!("response value: {:?}", response_value);
+    //     info!("response value: {:?}", response_value);
 
-        if !response_value.balances.coins.is_empty() {
-            break;
-        } else {
-            std::thread::sleep(Duration::from_secs(5));
-        }
-    }
+    //     if !response_value.balances.coins.is_empty() {
+    //         break;
+    //     } else {
+    //         std::thread::sleep(Duration::from_secs(5));
+    //     }
+    // }
 
     std::thread::sleep(Duration::from_secs(3));
+    let addr = "osmo1hj5fveer5cjtn4wd6wstzugjfdxzl0xpwhpz63";
+
+    let converted_addr_bytes = decode_and_convert(&addr).unwrap();
+
+    let balance_key = create_account_denom_balance_key(converted_addr_bytes, "uosmo").unwrap();
+
     let kvq_registration_response = register_kvq_balances_query(
         &test_ctx,
         icq_test_lib.address.to_string(),
         OSMOSIS_CHAIN_NAME.to_string(),
-        OSMOSIS_CHAIN_ADMIN_ADDR.to_string(),
+        BANK_STORE_KEY.to_string(),
+        balance_key,
     )?;
 
     info!(
@@ -119,7 +142,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         kvq_registration_response
     );
 
-    for _ in 0..10 {
+    for _ in 0..2 {
         let response_value = query_catchall_logs(&test_ctx, icq_test_lib.address.to_string())?;
         info!("catchall logs: {:?}", response_value);
 
@@ -143,170 +166,4 @@ fn main() -> Result<(), Box<dyn Error>> {
     }
 
     Ok(())
-}
-
-fn try_parse_storage_value(storage_value: &StorageValue) -> Value {
-    let mut map = serde_json::Map::new();
-
-    // Add storage prefix
-    map.insert(
-        "storage_prefix".to_string(),
-        Value::String(storage_value.storage_prefix.clone()),
-    );
-
-    // Try UTF-8 string interpretation
-    if let Ok(key_str) = String::from_utf8(storage_value.key.to_vec()) {
-        map.insert("key_utf8".to_string(), Value::String(key_str));
-    }
-
-    if let Ok(value_str) = String::from_utf8(storage_value.value.to_vec()) {
-        map.insert("value_utf8".to_string(), Value::String(value_str));
-    }
-
-    // Try JSON interpretation
-    if let Ok(value_str) = String::from_utf8(storage_value.value.to_vec()) {
-        if let Ok(json_value) = serde_json::from_str(&value_str) {
-            map.insert("value_json".to_string(), json_value);
-        }
-    }
-
-    if let Ok(value_str) = String::from_utf8(storage_value.key.to_vec()) {
-        if let Ok(json_value) = serde_json::from_str(&value_str) {
-            map.insert("key_json".to_string(), json_value);
-        }
-    }
-
-    // Convert raw bytes to base64
-    map.insert(
-        "key".to_string(),
-        Value::String(storage_value.key.to_string()),
-    );
-
-    Value::Object(map)
-}
-
-fn register_kvq_balances_query(
-    test_ctx: &TestContext,
-    icq_lib: String,
-    domain: String,
-    addr: String,
-) -> Result<TransactionResponse, LocalError> {
-    info!("registering ICQ KV query on domain {domain}...");
-
-    let register_kvq_msg = ExecuteMsg::RegisterKeyValueQuery {
-        connection_id: test_ctx
-            .get_connections()
-            .src(NEUTRON_CHAIN_NAME)
-            .dest(&domain)
-            .get(),
-        update_period: 5,
-        key: addr,
-    };
-
-    contract_execute(
-        test_ctx
-            .get_request_builder()
-            .get_request_builder(NEUTRON_CHAIN_NAME),
-        &icq_lib,
-        DEFAULT_KEY,
-        &serde_json::to_string(&register_kvq_msg)
-            .map_err(|e| LocalError::Custom { msg: e.to_string() })?,
-        "--amount 1000000untrn --gas 50000000",
-    )
-}
-
-pub fn register_icq_balances_query(
-    test_ctx: &TestContext,
-    icq_lib: String,
-    domain: String,
-    addr: String,
-    denoms: Vec<String>,
-) -> Result<TransactionResponse, LocalError> {
-    info!("registering ICQ balances query on domain {domain} for {addr}...");
-
-    let register_icq_msg = ExecuteMsg::RegisterBalancesQuery {
-        connection_id: test_ctx
-            .get_connections()
-            .src(NEUTRON_CHAIN_NAME)
-            .dest(&domain)
-            .get(),
-        update_period: 5,
-        addr,
-        denoms,
-    };
-
-    contract_execute(
-        test_ctx
-            .get_request_builder()
-            .get_request_builder(NEUTRON_CHAIN_NAME),
-        &icq_lib,
-        DEFAULT_KEY,
-        &serde_json::to_string(&register_icq_msg)
-            .map_err(|e| LocalError::Custom { msg: e.to_string() })?,
-        "--amount 1000000untrn --gas 50000000",
-    )
-}
-
-pub fn query_balance_query_id(
-    test_ctx: &TestContext,
-    icq_lib: String,
-    query_id: u64,
-) -> Result<neutron_sdk::interchain_queries::v047::queries::BalanceResponse, LocalError> {
-    let query_response = contract_query(
-        test_ctx
-            .get_request_builder()
-            .get_request_builder(NEUTRON_CHAIN_NAME),
-        &icq_lib,
-        &serde_json::to_string(&QueryMsg::Balance { query_id })
-            .map_err(|e| LocalError::Custom { msg: e.to_string() })?,
-    )["data"]
-        .clone();
-
-    let balance_response: neutron_sdk::interchain_queries::v047::queries::BalanceResponse =
-        serde_json::from_value(query_response).unwrap();
-
-    info!("balance query response: {:?}", balance_response);
-
-    Ok(balance_response)
-}
-
-pub fn query_raw_result(
-    test_ctx: &TestContext,
-    icq_lib: String,
-    query_id: u64,
-) -> Result<InterchainQueryResult, LocalError> {
-    let query_response = contract_query(
-        test_ctx
-            .get_request_builder()
-            .get_request_builder(NEUTRON_CHAIN_NAME),
-        &icq_lib,
-        &serde_json::to_string(&QueryMsg::RawIcqResult { id: query_id })
-            .map_err(|e| LocalError::Custom { msg: e.to_string() })?,
-    )["data"]
-        .clone();
-
-    let icq_result: InterchainQueryResult = serde_json::from_value(query_response).unwrap();
-
-    info!("raw icq result: {:?}", icq_result);
-
-    Ok(icq_result)
-}
-
-pub fn query_catchall_logs(
-    test_ctx: &TestContext,
-    icq_lib: String,
-) -> Result<Vec<(String, String)>, LocalError> {
-    let query_response = contract_query(
-        test_ctx
-            .get_request_builder()
-            .get_request_builder(NEUTRON_CHAIN_NAME),
-        &icq_lib,
-        &serde_json::to_string(&QueryMsg::Catchall {})
-            .map_err(|e| LocalError::Custom { msg: e.to_string() })?,
-    )["data"]
-        .clone();
-
-    let resp: Vec<(String, String)> = serde_json::from_value(query_response).unwrap();
-
-    Ok(resp)
 }
